better targeting for checkbox/radio inputs
Instead of just looking at generic "inputs", we take a look at and
perform various actions based on the input type:

- checkbox: click, blur
- radio: click, blur
- select: click

For checkboxes/radios we blur immediately following the click so that we
can perform actions on the page again.

With select elements we leave the focus on the element so that a user
can use the arrow keys to select something from the dropdown.